refactor: unify Last.fm client
